requirements.txt
python-telegram-bot==13.7
flask==2.0.1
flask-cors==3.0.10 